requirements.txt
fastapi
uvicorn[standard]
mongo
databases